chat_server/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
dataclasses-json==0.5.14
docarray==0.32.1
exceptiongroup==1.1.3
faiss-cpu==1.7.4
filelock==3.12.3
Flask==3.0.0
Flask-Cors==4.0.0
flatbuffers==23.5.26
frozenlist==1.4.0
fsspec==2023.9.0
google-api-core==2.11.1
google-api-python-client==2.98.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.60.0
greenlet==2.0.2
h11==0.14.0
hnswlib==0.7.0
html2text==2020.1.16
httplib2==0.22.0
httptools==0.6.0
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
langchain==0.0.281
langsmith==0.0.33
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.5
numpy==1.25.2
onnxruntime==1.15.1
openai==0.28.0
orjson==3.9.7
overrides==7.4.0
packaging==23.1
posthog==3.0.2
protobuf==4.24.3
pulsar-client==3.3.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.12
pydantic_core==2.6.3
Pygments==2.16.1
pyparsing==3.1.1
PyPika==0.48.9
pyreadline3==3.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
rich==13.6.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.20
starlette==0.27.0
sympy==1.12
tenacity==8.2.3
tiktoken==0.4.0
tokenizers==0.14.0
tqdm==4.66.1
types-requests==2.28.11.6
types-urllib3==1.26.25.14
typing-inspect==0.9.0
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.16
watchfiles==0.20.0
websockets==11.0.3
Werkzeug==3.0.0
yarl==1.9.2
zipp==3.16.2
